requirements-gemma3-extra.txt
accelerate==1.6.0
antlr4-python3-runtime==4.11.0
chardet==5.2.0
colorama==0.4.6
DataProperty==1.1.0
datasets==3.5.0
docker-pycreds==0.4.0
evaluate==0.4.3
fire==0.7.0
gitdb==4.0.12
GitPython==3.1.44
huggingface-hub==0.30.2
jsonlines==4.0.0
-e ./lm-evaluation-harness
lxml==5.3.2
mbstrdecoder==1.1.4
multiprocess==0.70.17
nltk==3.9.1
numexpr==2.10.2
pathvalidate==3.2.3
-e ./peft
pipdeptree==2.23.1
portalocker==3.1.1
pytablewriter==1.2.1
responses==0.18.0
rouge_score==0.1.2
sacrebleu==2.5.1
sentencepiece==0.2.0
sentry-sdk==2.26.1
setproctitle==1.3.5
smmap==5.0.2
sqlitedict==2.1.0
tabledata==1.3.4
tcolorpy==0.1.7
termcolor==3.0.1
tokenizers==0.21.1
tqdm-multiprocess==0.0.11
transformers==4.50.3
typepy==1.3.4
wandb==0.19.9
word2number==1.1
xxhash==3.5.0
zstandard==0.23.0
